requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
langsmith==0.0.69
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.27.2